requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
idna==3.4
PyYAML==6.0
requests==2.28.1
urllib3==1.26.12
